Bluetooth: Mesh: Add return value for opcode callback
```
3.7.3.4 Message error procedure
When receiving a message that is not understood by an element, it shall
ignore the message.
Note: A message can be falsely identified as a valid message, passing
the NetMIC and TransMIC authentication using a known network key and
application key even though that message was sent using different keys.
The decryption of that message using the wrong keys would result in a
message that is not understood by the element. The probability of such a
situation occurring is small but not insignificant.
A message that is not understood includes messages that have one or more
of the following conditions:
• The application opcode is unknown by the receiving element.
• The access message size for the application opcode is incorrect.
• The application parameters contain values that are currently
Prohibited.
Note: An element that sends an acknowledged message that is not
understood by a peer node will not receive any response message.
```

Mesh Profile 1.1 Specification introduces Opcodes Aggregator feature:
```
4.4.19.2.1 Message Request List Processing procedure

Each access message shall be processed by the identified model by
preforming the following steps:

[...]
3. When an access message is not successfully processed by the
identified model (i.e., it results in any of the error conditions in
Table 4.360), processing of the message request list shall stop, and the
procedure shall be completed by returning the current message results
list and the Status set to the error condition as defined in Table 4.360.

[...]
Table 4.360: Error conditions for the message results list

An access message has a valid opcode but is not understood by the
identified model (see Section 3.7.3.4) | MessageNotUnderstood
```

Thus we need a way for a model to return some status information after
processing an access message so that we can set an appropriate status
response when using Opcodes Aggregator. In normal operation them message
will just be ignored.

Signed-off-by: Michał Narajowski <michal.narajowski@codecoup.pl>