Bluetooth: Mesh: Refactor Mesh Model Extensions
The existing extension tree does not support all the features that are
defined by the specification (e.g. multiple parents).

To enable full support for Model Extensions this patch introduces graph
representation in a form of edge list. This representation allows for
fine-grained control over memory usage. The downside is that traversing
the graph is sub-optimal computationally.

Signed-off-by: Michał Narajowski <michal.narajowski@codecoup.pl>